==> Sup_Note5/R/6_Table10.R <==
library(tidyverse)
Bench1 <- readRDS("data/IntestinalBenchCell.rds")%>%  spread(methods, value) %>%  rename(`Reference Data` = data)
Bench1A <- Bench1 %>%  filter(cell_type != "Rejection") 
Bench1B <- Bench1 %>%  filter(cell_type == "Rejection")%>%  filter(metrics == "Recall")
Bench1B <- Bench1B[,-(2:3)] %>%  gather("methods", "rejection rate",-1)
Bench2 <- readRDS("data/IntestinalBenchOverall.rds")%>%  spread(methods, value) %>%  rename(`Reference Data` = data)
xlsx::write.xlsx(Bench1A, file = "../FinalTable/SupTable5.xlsx", sheetName = "Cell_Population", append = F)
xlsx::write.xlsx(Bench1B, file = "../FinalTable/SupTable5.xlsx", sheetName = "Rejection", append = T)
xlsx::write.xlsx(Bench2, file = "../FinalTable/SupTable5.xlsx", sheetName = "Overall", append = T)
